Removed OnlineWardleyMaps as basis for this standard
Removed references to OnlineWardleyMaps as the basis for this standard.  While it makes sense to learn from implementations, we should NOT base a standard description language on a single tool.  Instead, we should base it on the standard described in Simon's book, subsequent revisions, and user's needs (including OWM's).

Added statement on user needs.

Marked this as DRAFT so it can be merged to main without consequences